requirements.txt
boto3
opencv-python-headless
numpy
imutils
Pillow
boto3
facenet-pytorch